+++ src/Symfony/Component/DependencyInjection/Compiler/PriorityTaggedServiceTrait.php
@@ -50,6 +50,7 @@ trait PriorityTaggedServiceTrait
             $tagName = $tagName->getTag();
         }
 
+        $parameterBag = $container->getParameterBag();
         $i = 0;
         $services = [];
 
@@ -81,8 +82,9 @@ trait PriorityTaggedServiceTrait
                 }
 
                 if (null !== $indexAttribute && isset($attribute[$indexAttribute])) {
-                    $index = $attribute[$indexAttribute];
-                } elseif (null === $defaultIndex && $defaultPriorityMethod && $class) {
+                    $index = $parameterBag->resolveValue($attribute[$indexAttribute]);
+                }
+                if (null === $index && null === $defaultIndex && $defaultPriorityMethod && $class) {
                     $defaultIndex = PriorityTaggedServiceUtil::getDefault($container, $serviceId, $class, $defaultIndexMethod ?? 'getDefaultName', $tagName, $indexAttribute, $checkTaggedItem);
                 }
                 $decorated = $definition->getTag('container.decorator')[0]['id'] ?? null;