+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Add a `ContainerBuilder::registerChild()` shortcut method for registering child definitions
  * Add support for `key-type` in `XmlFileLoader`
  * Enable non-empty parameters with `ParameterBag::cannotBeEmpty()` and `ContainerBuilder::parameterCannotBeEmpty()` methods
+ * Resolve parameters found in index attribute of service tags
 
 7.1
 ---